remove ome_utils
